Starting forward slash issue on Laravel 4.1
Spotted an issue with latest of Laravel (maybe since version 4.1.6? not sure) when generating the links in the language bar caused by a '/' not present anymore in the $path variable (and therefore no str_replace is happening creating links like 'en/acerca').
In a DRY way I choose to add a starting forward slash when $path doesn't start with it.