Added a new getCurrentLocaleNative function
Added a new ```getCurrentLocaleNative``` function, there are times when you want to display the current locale in the user's native language.  If a user selects Deutsch (German) in a drop down menu to charge their language to german, once the locale has been set and changed the site's language displays everything in German but the ````getCurrentLocaleName``` displays the word German not Deutsch.  So in cases like this use ```getCurrentLocaleNative``` instead of ```getCurrentLocaleName```.