Fix route caching on Laravel v5.8 (#626)
Laravel v5.8 has a bit different way of getting environment variables.
The `ROUTING_LOCALE` variable is probably not set at boot time so it will not be available later on in the app lifecycle since `getenv` function is not used directly.
This fix is just using `getenv` as a fallback. Another option would be to just use it directly and not as a fallback for `env`.