Fix hideDefaultLocale
A previous commit for caching broke
the functionality of `hideDefaultLocale`.

The purpose of `setLocale` is to return the
locale of the url (even if there is none).
This is necessary for ensuring that the
current requested route actually exists.
Therefore this behaviour was reestablished.

The same behaviour goes for the cached routes.
The file `bootstrap/cache/routes_**.php` imitates
the routes from `setLocale`. Therefore, `getCurrentLocale`
has to be dismissed. Also a nullable locale should
refere to `bootstrap/cache/routes.php` and not
an alternative.

Fixes #656
Fixes #661