Imrove description of useAcceptLanguageHeader and hideDefaultLocaleInURL
Many issues from the repository show that many people have a hard
time to understand how both configuration works.
Also the description of `hideDefaultLocaleInURL` was outdated in
the config file.

This commit is an attempt to be more clear about these configuration
setup and the behaviour of the associated middleware.