Improve config description and readme.md
Many issues show that it is difficult to understand
how `hideDefaultLocaleInURL` and `useAcceptLanguageHeader`
work, especially how the middleware will effect them
and if they are used in combination.
Also the description of hideDefaultLocaleInURL was outdated in
the config file.

This commit is an attempt to be more clear about these configuration
setup and the behaviour of the associated middleware.