Add an example with attributes for getURLFromRouteNameTranslated
This appears to be the most direct way to make links within a website that leverage the `routes.php` files. The lack of documentation for the use of the attributes took me way to long to look up. So here I am providing a simple addition to the documentation to help others (and myself) in the future.